+++ accounting/src/sdk/sdk.ts
@@ -75,9 +75,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.4.0";
-    genVersion = "2.188.1";
-    userAgent = "speakeasy-sdk/typescript 0.4.0 2.188.1 3.0.0 @speakeasy-sdks/accounting";
+    sdkVersion = "0.4.1";
+    genVersion = "2.194.1";
+    userAgent = "speakeasy-sdk/typescript 0.4.1 2.194.1 3.0.0 @speakeasy-sdks/accounting";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);